advancing with jwt, not finished but got the first part.

+++ app/auth.py
@@ -10,6 +10,7 @@ from .utils import data_encrypt
 import sqlalchemy as sa
 
 
+jwt = JWTManager()
 auth_bp= Blueprint('auth_bp', __name__)
 
 @login_required
@@ -18,12 +19,9 @@ def login():
     form = LoginForm(request.form)
     if current_user.is_authenticated:
         return redirect(url_for('main_bp.index'))
-
-        
-
+    
 
     if request.method == 'POST':
-        
         if form.validate_on_submit():
             user = db.session.scalar(
                     sa.select(User).where(User.username == form.username.data))
@@ -34,12 +32,12 @@ def login():
             next_page = request.args.get('next')
             if not next_page or urlsplit(next_page).netloc != '':
                 next_page = url_for('main_bp.index')
-                access_token = create_access_token(identity=user.id)
+                session['jwt'] = create_access_token(identity=user.id)      
                 #need to store the jwt with user id for further
                 #use. implement logic here.
-                return redirect(next_page)
+                return redirect(url_for('auth_bp.upload')), 200
+            
     return render_template('login.html', title='Sign In', form=LoginForm())
-
 
 @auth_bp.route('/register', methods=['GET', 'POST'])
 def register():
@@ -55,23 +53,32 @@ def register():
         db.session.add(user)
         db.session.commit()
         flash('Congrats! You are now a registered user!')
+        access_token = create_access_token(identity=user.id)
+        user_id = User.id
+        session[user_id] = access_token
         return redirect(url_for('auth_bp.login'))
     return render_template('register.html', title='Register', form=form)
 
 
-auth_bp.route('/upload', methods=['POST'])
+@auth_bp.route('/upload', methods=['GET', 'POST'])
 @jwt_required()
-def upload_encrypted_data():
-   with open(file_path, 'rb') as f:
-       file_content = f.read()
-
-   encrypted_data_dict = data_encrypt(file_content)
-    # logic to upload the file to ipfs
-
-   os.remove(file_path)
-
-   return jsonify({'message': 'File uploaded and encrypted successfully'}), 200
-
+def upload():
+    if request.method == 'POST':
+        user_id = User.id
+        print(user_id)
+        access_token =get_jwt_identity()
+        print(access_token)
+        if access_token:
+            with open(file_path, 'rb') as f:
+                file_content = f.read()
+                encrypted_data_dict = data_encrypt(file_content)
+                #logic to upload the file to ipfs
+                os.remove(file_path)
+                return jsonify({'message': 'File uploaded and encrypted successfully'}), 200
+        else:
+            return jsonify({'message' : 'Flie not uploaded. No JWT(security token) found.'}), 400
+    if request.method == 'GET':
+        return render_template('upload.html')
